requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-admin-interface==0.28.9
django-cleanup==9.0.0
django-colorfield==0.11.0
django-import-export==4.1.1
django-translation==1.2
pillow==10.4.0
python-slugify==8.0.4
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
